feat(frontend): add closet and outfit page feat(frontend): use remove.bg api to remove clothes background

+++ frontend/pages/home/SettingsPage.js
@@ -1,6 +1,12 @@
-import { View, Button, Text, StyleSheet } from 'react-native';
+import { View, TouchableOpacity, Text, StyleSheet, Alert } from 'react-native';
 import { useAuth, useUser } from '@clerk/clerk-expo';
+import UserModel from '../../models/UserModel';
 
+const Button = ({ onPress, title, backgroundColor, textColor }) => (
+    <TouchableOpacity onPress={onPress} style={[styles.button, { backgroundColor: backgroundColor }]}>
+        <Text style={[styles.text, { color: textColor }]}>{title}</Text>
+    </TouchableOpacity>
+);
 
 const SignOut = () => {
     const { isLoaded, signOut } = useAuth();
@@ -8,14 +14,7 @@ const SignOut = () => {
         return null;
     }
     return (
-        <View>
-            <Button
-                title="Sign Out"
-                onPress={() => {
-                    signOut();
-                }}
-            />
-        </View>
+        <Button onPress={signOut} title="登出" backgroundColor="#000" textColor="#fff" />
     );
 };
 
@@ -24,32 +23,50 @@ const DeleteAccount = () => {
     const { _, isSignedIn, user } = useUser();
 
     async function deleteAccount() {
-        // const token = await getToken();
-        await UserModel.deleteUser(user.id);
-        signOut();
+        Alert.alert('刪除帳號', '真的要刪除嗎?', [
+            {
+                text: '取消',
+                onPress: () => console.log('Cancel Pressed'),
+                style: 'cancel',
+            },
+            {
+                text: '確認', onPress: async () => {
+                    await UserModel.deleteUser(user.id);
+                    signOut();
+                }
+            }
+        ]);
+
     }
     if (!isLoaded) {
         return null;
     }
     return (
-        <View>
-            <Button
-                title="Delete Account"
-                onPress={() => {
-                    deleteAccount();
-                }}
-            />
-        </View>
+        <Button onPress={deleteAccount} title="刪除帳號" backgroundColor="#FF0000" textColor="#fff" />
     );
 
 }
 
-export default function SettingsPage() {
+const UpdateUser = ({ user, navigation }) => {
+    async function gotoUpdate() {
+        navigation.navigate("UpdateUser", { user: user });
+    }
+    return (
+        <Button onPress={gotoUpdate} title="更新個人資料" backgroundColor="#000" textColor="#fff" />
+    );
+}
+
+export default function SettingsPage({ route, navigation }) {
+    const user = route.params.user;
+
     return (
         <View style={styles.container}>
-            <Text>Settings Page</Text>
-            <SignOut />
-            <DeleteAccount />
+            <View style={styles.buttonContainer}>
+                <UpdateUser navigation={navigation} user={user} />
+                <SignOut />
+                <DeleteAccount />
+            </View>
+
         </View>
     )
 }
@@ -59,5 +76,18 @@ const styles = StyleSheet.create({
         flex: 1,
         alignItems: "center",
         justifyContent: "center",
+    },
+    buttonContainer: {
+        width: "80%",
+        height: "80%",
+        justifyContent: "center",
+    },
+    button: {
+        alignItems: "center",
+        justifyContent: "center",
+        width: "100%",
+        height: 50,
+        borderRadius: 10,
+        margin: 5,
     }
 })
